Scottx611x/handle lack of galaxy input step inputs (#2590)
* Handle case where Galaxy now doesn't include information about an input step's inputs

* Remove unused constants

* Move `create_expanded_workflow_graph` into tool_manager.utils

* Remove blank file

* Smiplify import
